Setting up local storage to save vars
