+++ ClientMicroBlog/public_html/login.html
@@ -9,9 +9,7 @@
         <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
     
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-        <script src="scripts/LOGIN.js"></script>
         <script src="scripts/APP.js"></script>
-        <script src="scripts/USER.js"></script>
     </head>
     <body>
         <div class="container">
@@ -39,7 +37,7 @@
         
         <script>
             $(document).ready(function() {
-                $("#btn_login").on("click", LOGIN.doLogin);
+                $("#btn_login").on("click", APP.doLogin);
             });
         </script>
     </body>
