--- ClientMicroBlog/public_html/scripts/USER.js
@@ -1,12 +1,0 @@
-
-
-/**
- * @author gfurri20
- */
-var USER =
-{
-    USERNAME: null,
-    JWT_TOKEN: null,
-    HEADER_STRING: "Authorization",
-    TOKEN_PREFIX: "Bearer "
-};
